requirements.txt
# Polymarket API client — V2 (py-clob-client stops working after April 28, 2026)
py-clob-client-v2
 
# HTTP client for market discovery and price fetching
httpx==0.27.0
 
# Environment configuration
python-dotenv==1.0.1
 